rename app to landing pages in prep for a second entry file for embed only
